@remotion/studio-codemods: Share Solid insertion
