+++ packages/create-video/src/list-of-remotion-packages.ts
@@ -70,6 +70,7 @@ export const listOfRemotionPackages = [
 	'@remotion/serverless-client',
 	'@remotion/skills',
 	'@remotion/skills-evals',
+	'@remotion/studio-codemods',
 	'@remotion/studio-server',
 	'@remotion/studio-shared',
 	'@remotion/studio',
