+++ packages/studio-codemods/README.md
@@ -1,0 +1,5 @@
+# @remotion/studio-codemods
+
+Shared codemods used internally by the Remotion Studio server and Browser Studio.
+
+This package is an internal implementation detail of Remotion Studio.
