+++ packages/studio-codemods/package.json
@@ -1,17 +1,15 @@
 {
-	"repository": {
-		"url": "https://github.com/remotion-dev/remotion/tree/main/packages/browser-studio"
-	},
-	"name": "@remotion/browser-studio",
-	"private": true,
+	"name": "@remotion/studio-codemods",
 	"version": "4.0.500",
-	"description": "Run Remotion Studio in the browser",
-	"main": "dist/esm/index.mjs",
+	"description": "Shared codemods for Remotion Studio",
+	"main": "dist/index.cjs",
 	"types": "dist/index.d.ts",
 	"module": "dist/esm/index.mjs",
+	"repository": {
+		"url": "https://github.com/remotion-dev/remotion/tree/main/packages/studio-codemods"
+	},
 	"type": "module",
 	"scripts": {
-		"dev": "bun src/dev/server.ts",
 		"test": "bun test src/test",
 		"formatting": "oxfmt src --check",
 		"format": "oxfmt src",
@@ -21,42 +19,34 @@
 	},
 	"author": "Jonny Burger <jonny@remotion.dev>",
 	"contributors": [],
-	"license": "Remotion License",
+	"license": "MIT",
 	"bugs": {
 		"url": "https://github.com/remotion-dev/remotion/issues"
 	},
 	"dependencies": {
-		"@remotion/studio-codemods": "workspace:*",
-		"@remotion/studio-shared": "workspace:*",
-		"@rspack/browser": "2.1.1"
+		"@babel/parser": "7.24.1",
+		"@remotion/studio-shared": "workspace:*"
 	},
 	"exports": {
 		".": {
 			"types": "./dist/index.d.ts",
+			"require": "./dist/index.cjs",
 			"module": "./dist/esm/index.mjs",
 			"import": "./dist/esm/index.mjs"
-		},
-		"./browser-studio-worker": {
-			"types": "./dist/browser-studio-worker.d.ts",
-			"module": "./dist/esm/browser-studio-worker.mjs",
-			"import": "./dist/esm/browser-studio-worker.mjs"
 		},
 		"./package.json": "./package.json"
 	},
 	"devDependencies": {
-		"@remotion/studio": "workspace:*",
 		"@remotion/eslint-config-internal": "workspace:*",
-		"@types/react": "catalog:",
-		"@types/react-dom": "catalog:",
 		"@typescript/native-preview": "catalog:",
-		"eslint": "catalog:",
-		"react": "catalog:",
-		"react-dom": "catalog:"
+		"eslint": "catalog:"
 	},
 	"keywords": [
 		"remotion",
 		"studio",
-		"rspack",
-		"browser"
-	]
+		"codemod"
+	],
+	"publishConfig": {
+		"access": "public"
+	}
 }
