+++ packages/studio-server/package.json
@@ -35,6 +35,7 @@
 		"recast": "0.23.11",
 		"@remotion/bundler": "workspace:*",
 		"@remotion/renderer": "workspace:*",
+		"@remotion/studio-codemods": "workspace:*",
 		"@remotion/studio-shared": "workspace:*",
 		"memfs": "3.4.3",
 		"open": "8.4.2"
