+++ packages/studio-shared/src/package-info.ts
@@ -72,6 +72,7 @@ const allPackages = [
 	'serverless-client',
 	'skills',
 	'skills-evals',
+	'studio-codemods',
 	'studio-server',
 	'studio-shared',
 	'studio',
@@ -162,6 +163,7 @@ export const descriptions: {[key in Pkgs]: string | null} = {
 	bundler: 'Bundle Remotion compositions using Webpack',
 	'browser-studio': 'Run Remotion Studio in the browser',
 	'canvas-capture': 'Capture HTML-in-canvas content as a video',
+	'studio-codemods': 'Shared codemods for Remotion Studio',
 	'studio-server': 'Run a Remotion Studio with a server backend',
 	'install-whisper-cpp': 'Helpers for installing and using Whisper.cpp',
 	'whisper-web': 'Helpers for using Whisper.cpp in browser using WASM',
@@ -328,6 +330,7 @@ export const installableMap: {[key in Pkgs]: boolean} = {
 	'promo-pages': false,
 	streaming: false,
 	serverless: false,
+	'studio-codemods': false,
 	'studio-server': false,
 	'studio-shared': false,
 	studio: true,
@@ -385,6 +388,7 @@ export const apiDocs: {[key in Pkgs]: string | null} = {
 	bundler: 'https://www.remotion.dev/docs/bundler',
 	'browser-studio': null,
 	'canvas-capture': null,
+	'studio-codemods': null,
 	'lambda-client': null,
 	'serverless-client': null,
 	'studio-server': null,
