+++ tsconfig.json
@@ -199,6 +199,9 @@
 			"path": "./packages/studio"
 		},
 		{
+			"path": "./packages/studio-codemods"
+		},
+		{
 			"path": "./packages/studio-server"
 		},
 		{
